heroku/heroku_save/requirements.txt



dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
discord.py==1.2.5
python-dotenv==0.10.2
requests==2.7.0
bs4==0.0.1
googletrans==2.4.0
